+++ src/benchmarker/__init__.py
@@ -1,3 +1,7 @@
+""" 
+The becckmarker for
+"""
+
 from .bm import BenchMark