+++ CHANGELOG.md
@@ -33,4 +33,27 @@ type: beta
 
 - Renamed the command base class from `HList` to `Comm`. **(tag: name_scheme)**
 
-- Added `e2eeftp.server.server.E2EEFTPRequestHandler.update_command_handlers()` to update the `e2eeftp.server.server.E2EEFTPRequestHandler.command_handlers: dict[str, str]` each time the server is initialized. **(tag: feature)**
+- Added `method:e2eeftp.server.server.E2EEFTPRequestHandler.update_command_handlers() -> None` to update the `e2eeftp.server.server.E2EEFTPRequestHandler.command_handlers: dict[str, str]` each time the client sends an request. **(tag: feature)**
+
+- Changed `class:e2eeftp.client.client.e2eeftpClient` to also work with `with..as` statement. **(tag: feature_change)**
+    ```python
+    # syntax: 
+    from e2eeftp import e2eeftpClient
+    
+    with e2eeftpClient(host='127.0.0.1', port=8080) as client:
+        client.send("file.jpg")    # Testing send request
+        client.get('file.jpg')     # Testing get request
+        print(client.list_files()) # Testing list request
+        print(client.hlist())      # Testing hlist request
+    ```
+
+- Added `method:e2eeftp.server.server.E2EEFTPRequestHandler.setup() -> None` to run the setup script `method:e2eeftp.server.server.E2EEFTPRequestHandler.update_command_handlers() -> None`. **(tag: feature)**
+
+- Added docstrings:
+    * `example._server.py`
+    * `example.client_cli.py`
+    * `example.client.py`
+    * `example.server.py`
+    * `src.benchmarker.__init__.py`
+    * `src.benchmarker.__init__.py`
+    * `src.auth.e2ee.py`